pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.45.0",
    "google-auth>=2.49.2",
    "google-cloud-aiplatform>=1.71.1",
    "instructor>=1.15.1",
    "ipython>=9.12.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "nltk>=3.9.4",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.4.0",
    "supermemory>=3.33.0",
    "qdrant-client>=1.17.0",
    "vertexai>=1.71.1",
]
